Added some more comments and smaller refactorings
