+++ app/src/app/components/transformation-view/transformation-view.component.html
@@ -26,7 +26,7 @@
         </li>
         <li *ngIf="transformationDone" class="nav-item">
             <a class="nav-link" href="javascript:void(0)" [class.active]="state === 'outputs'"
-               (click)="changeState()">Outputs</a>
+               (click)="openOutputTab()">Outputs</a>
         </li>
     </ul>
     <div *ngIf="state === 'inputs'">
